Seed QC check

+++ Rfam/Scripts/jiffies/rewrite_seed_with_rf.pl
@@ -30,30 +30,50 @@
 
 use strict;
 use warnings;
-#use Cwd;
-#use Getopt::Long;
+use Getopt::Long;
 use File::Copy;
-#use Data::Printer;
-#use Carp;
 
 use Bio::Rfam::Config;
 use Bio::Rfam::FamilyIO;
-#use Bio::Rfam::Family::MSA;
 use Bio::Rfam::Infernal;
 use Bio::Rfam::Utils;
 
 use Bio::Easel::MSA;
-#use Bio::Easel::SqFile;
-#use Bio::Easel::Random;
+
+use Getopt::Long;
+
+my $usage  = "rewrite_seed_with_rf.pl: updates SEED to follow infernal RF conventions,\n";
+$usage .= "                         preserving pseudoknots, and verifies new SEED would\n";
+$usage .= "                         create an identical CM to existing CM. Other files\n";
+$usage .= "                         (DESC, REVTLBOUT, SCORES, TBLOUT, SEEDTBLOUT, SEEDSCORES)\n";
+$usage .= "                         are 'touch'ed to update their timestamps (this is ok since\n";
+$usage .= "                         we verified CM is unchanged). This script should be called\n";
+$usage .= "                         from within a family directory (requires other files besides SEED).\n\n";
+$usage .= "Usage: perl rewrite_seed_with_rf.pl [OPTIONS] SEED\n\n";
+$usage .= "\tOPTIONS:\n";
+$usage .= "\t--keep:  leave all intermediate output files instead of deleting them\n";
+$usage .= "\t--quiet: do not output per-position SS_cons differences between input SEED and newly created SEED\n";
+
+my $do_keep  = 0;
+my $do_quiet = 0;
+&GetOptions( "keep"  => \$do_keep,
+             "quiet" => \$do_quiet);
+
+if(scalar(@ARGV) != 1) { die $usage; }
+my ($in_seed) = @ARGV;
+
+if($in_seed ne "SEED") { die $usage; }
 
 # required files in three 'tiers'. Files in tier 1 must be younger
 # than those in tiers 2 and 3, and files in tier 2 must be younger
 # than those in tier 3.
 my @reqdFiles_t1 = ("SEED");
 my @reqdFiles_t2 = ("CM");
-my @reqdFiles_t3 = ("DESC", "outlist", "outlist.pdf", "REVTBLOUT", "SCORES", "species",
-                    "species.pdf", "taxinfo", "TBLOUT");
-my @nonreqdFiles = ("revoutlist", "revspecies"); # these will only exist if there were >0 reverse hits
+#my @reqdFiles_t3 = ("DESC", "outlist", "outlist.pdf", "REVTBLOUT", "SCORES", "species",
+#                    "species.pdf", "taxinfo", "TBLOUT");
+#my @nonreqdFiles = ("revoutlist", "revspecies"); # these will only exist if there were >0 reverse hits
+my @reqdFiles_t3 = ("DESC", "REVTBLOUT", "SCORES", "TBLOUT", "SEEDTBLOUT", "SEEDSCORES");
+my @nonreqdFiles = ("outlist", "outlist.pdf", "species", "species.pdf", "taxinfo", "revoutlist", "revspecies"); # these will only exist if rfsearch.pl/rfmake.pl has been run
 
 foreach my $file (@reqdFiles_t1, @reqdFiles_t2, @reqdFiles_t3) { 
   if(! -e $file) { die "ERROR, required file $file does not exist"; }
@@ -151,7 +171,7 @@ foreach $file1 (@reqdFiles_t2) {
 #    matches the one in the CM file, and save the output alignment
 #    as SEED.1, we'll use this to map RF annotation to the original
 #    SEED.
-verify_cm_built_from_seed($config, "CM", "SEED", "SEED.1");
+verify_cm_built_from_seed($config, "CM", "SEED", "SEED.1", $do_keep);
 
 # B. verify the COM cmbuild line matches the passed in $desc_opts
 #    which were read from the DESC file.
@@ -176,8 +196,11 @@ if($desc_buildopts ne $cm_buildopts) { die "ERROR cmbuild option/command read fr
 # NEW WAY: use cmalign --mapali output alignment, which preserves the input alignment
 # (the original seed in this case) exactly (it never doctors parsetrees (yes, HMMER
 # works the same way)).
+
+# if --checkonly used, 
+
 if(-e "SEED") { copy("SEED", "SEED.0"); }
-add_rf_and_ss_cons_given_cmalign_mapali_output("SEED", "SEED.1", "SEED.2", $orig_cm->{cmHeader}->{clen});
+my $diff_str = add_rf_and_ss_cons_given_cmalign_mapali_output("SEED", "SEED.1", "SEED.2", $orig_cm->{cmHeader}->{clen}, $do_quiet);
 
 ##############################################################
 # 4. Builds even newer CM from even newer SEED and verifies 
@@ -192,8 +215,9 @@ if(-e "CM")   { copy("CM"  , "CM.0"); }
 
 Bio::Rfam::Infernal::cmbuild_wrapper($config, "$buildopts", "CM.1", "SEED.2", $outfile);
 verify_two_cms_have_identical_parameters("CM.1", "CM", 0.00000001);
-unlink $outfile;
-
+if(! $do_keep) { 
+  unlink $outfile;
+}
 
 ##############################################################
 # 5. Adds E-value stats from original CM to even newer CM.
@@ -207,7 +231,7 @@ else          { die "ERROR, CM.2 does not exist"; }
 if(-e "SEED.2") { copy("SEED.2", "SEED"); }
 else          { die "ERROR, SEED.2 does not exist"; }
 
-verify_cm_built_from_seed($config, "CM", "SEED");
+verify_cm_built_from_seed($config, "CM", "SEED", undef, $do_keep);
 # now CM and SEED are the updated versions
 
 ##############################################################
@@ -219,20 +243,28 @@ my $file;
 foreach $file (@reqdFiles_t1, @reqdFiles_t2, @reqdFiles_t3) { 
   if(! -e $file) { die "ERROR file $file does not exist"; }
   system("touch $file");
-  sleep(0.1);
+  sleep(1);
 }
 
 foreach $file (@nonreqdFiles) { 
   if(-e $file) { 
     system("touch $file");
   }
-  sleep(0.1);
+  sleep(1);
 }
 
 # clean up
-foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2") { 
-  if(-e $file) { unlink $file; }
+if(! $do_keep) { 
+  foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2") { 
+    if(-e $file) { unlink $file; }
+  }
 }
+
+print "New      CM and seed alignment saved as CM and SEED\n";
+print "Original CM and seed alignment saved as CM.0 and SEED.0\n";
+print "CM parameters verified to be identical in files CM and CM.0\n";
+print $diff_str . "\n";
+
 
 ###############
 # SUBROUTINES #
@@ -243,15 +275,18 @@ foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2") {
 # output from cmalign --mapali ($cmalign_mapali_infile) with a single extra sequence other
 # than the original SEED, add RF and SS_cons annotation to the original SEED alignment 
 # and save it as $outfile.
+#
+# Returns possibly multi-line string reporting per-position changes in SS_cons
+# or "SS_cons unchanged" if it does not change.
 sub add_rf_and_ss_cons_given_cmalign_mapali_output {
-
-  my ($orig_infile, $cmalign_mapali_infile, $outfile, $clen) = @_;
+  my ($orig_infile, $cmalign_mapali_infile, $outfile, $clen, $do_quiet) = @_;
 
   # read in original seed
   my $orig_seed = Bio::Easel::MSA->new({
     fileLocation => $orig_infile,
     forceText    => 1
-      });
+                                       });
+  my $save_orig_ss_cons = $orig_seed->get_ss_cons();
   
   # read in new seed plus consensus sequence aligned (created with cmalign --mapali SEED CM c.fa)
   my $tmp_seed  = Bio::Easel::MSA->new({
@@ -266,7 +301,7 @@ sub add_rf_and_ss_cons_given_cmalign_mapali_output {
   for($i = $orig_seed->nseq(); $i < $tmp_seed->nseq;  $i++) { $usemeA[$i] = 0; }
   my $new_seed = $tmp_seed->sequence_subset(\@usemeA);
   undef $tmp_seed;
-  
+
   # verify sequences are in the same order in each MSA
   for(my $i = 0; $i < $orig_seed->nseq(); $i++) { 
     if($orig_seed->get_sqname($i) ne $new_seed->get_sqname($i)) { 
@@ -328,10 +363,38 @@ sub add_rf_and_ss_cons_given_cmalign_mapali_output {
   $orig_seed->set_rf($orig_rf);
   $orig_seed->set_ss_cons($orig_ss_cons);
   $orig_seed->capitalize_based_on_rf();
+  $orig_seed->remove_gap_rf_basepairs(1); # 1: wussify SS_cons and change gap RF columns to '.' in SS_cons
+
+  # determine if SS_cons has changed, and construct return string
+  my $new_orig_ss_cons = $orig_seed->get_ss_cons();
+  my $ret_str = undef;
+  if($save_orig_ss_cons ne $new_orig_ss_cons) {
+    if(length($save_orig_ss_cons) == length($new_orig_ss_cons)) {
+      if(! $do_quiet) { 
+        $ret_str = "Between SEED and SEED.0, SS_cons changed, per-position differences:\n";
+        my @save_A = split("", $save_orig_ss_cons);
+        my @new_A  = split("", $new_orig_ss_cons);
+        for(my $z = 0; $z < scalar(@save_A); $z++) {
+          if($save_A[$z] ne $new_A[$z]) { 
+            $ret_str .= sprintf("\tposition %5d changed from %s to %s\n", $z, $save_A[$z], $new_A[$z]);
+          }
+        }
+      }
+      else {
+        $ret_str = "Between SEED and SEED.0, SS_cons changed.";
+      }
+    }
+    else { 
+      $ret_str = "Between SEED and SEED.0, SS_cons changed, lengths differ, compare SEED.0 and SEED to see differences.";
+    }
+  }
+  else {
+o    $ret_str = "Between SEED and SEED.0 SS_cons unchanged.";
+  }
   
   $orig_seed->write_msa($outfile);
   
-  return;
+  return $ret_str;
 }
 
 #######################################################################
@@ -553,9 +616,11 @@ sub add_evalue_and_com_lines_to_cm {
 #
 # If $stkfile is undefined, unlink it at end of function, otherwise
 # save output alignment as $stkfile and don't unlink it.
+#
+# If $do_keep: don't unlink any files
 sub verify_cm_built_from_seed { 
 
-  my ($config, $cmfile, $seedfile, $stkfile) = @_;
+  my ($config, $cmfile, $seedfile, $stkfile, $do_keep) = @_;
 
   my $unlink_stkfile;
 
@@ -574,12 +639,14 @@ sub verify_cm_built_from_seed {
   else { 
     $unlink_stkfile = 0;
   }
-  Bio::Rfam::Infernal::cmalign_wrapper($config, "", "", "--mapali $seedfile -o $stkfile", "CM", $fafile, undef, "", 1, 100, 1, 0, 0, "", 1, undef, 0);
+  Bio::Rfam::Infernal::cmalign_wrapper($config, "", "", "--mapstr --mapali $seedfile -o $stkfile", "CM", $fafile, undef, "", 1, 100, 1, 0, 0, "", 1, undef, 0);
   # we'll die if $seedfile wasn't used to build $cmfile
 
-  unlink $fafile;
-  if($unlink_stkfile) { 
-    unlink $stkfile;
+  if(! $do_keep) { 
+    unlink $fafile;
+    if($unlink_stkfile) { 
+      unlink $stkfile;
+    }
   }
 
   return;
